refactor: parameterize fitToStage, handleWheel, and touch handlers
Phase 2 of code deduplication:
- Add fitToStage() to base with options for maxScale, updateDraggable
- Add handleWheel() to base with hitLayer and updateDraggable options
- Add touch handlers to base (handleTouchMove, handleTouchEnd)
- Viewer: use base methods with updateDraggable and hitLayer options
- Editor: use base methods, override touch end for draggable
- Kiosk: use base fitToStage with custom scale limits
- Net reduction: ~40 lines
