+++ assets/js/seat_map/seat_map_base.js
@@ -322,6 +322,147 @@ export class SeatMapBase {
     throw new Error('getMaxScale must be implemented by subclass')
   }
   
+  // Fit content to viewport with optional scale limits
+  fitToStage(resetPosition, options = {}) {
+    const {
+      maxVisibleScale = 1.5,
+      minScaleOverride,
+      maxScaleOverride,
+      updateDraggable = false,
+      padding =60
+    } = options
+    
+    const bounds = this.getContentBounds()
+    const stageWidth = this.stage.width()
+    const stageHeight = this.stage.height()
+    
+    if (bounds.width === 0 || bounds.height === 0) return
+    
+    const scale = Math.min(
+      (stageWidth - padding) / bounds.width,
+      (stageHeight - padding) / bounds.height,
+      maxVisibleScale
+    )
+    const minScale = minScaleOverride ?? this.getMinScale()
+    const maxScale = maxScaleOverride ?? this.getMaxScale()
+    const clampedScale = clamp(scale, minScale, maxScale)
+    
+    if (resetPosition) {
+      this.stage.scale({ x: clampedScale, y: clampedScale })
+      const scaledWidth = bounds.width * clampedScale
+      const scaledHeight = bounds.height * clampedScale
+      this.stage.position({
+        x: (stageWidth - scaledWidth) / 2 - bounds.x * clampedScale,
+        y: (stageHeight - scaledHeight) / 2 - bounds.y * clampedScale
+      })
+    }
+    
+    if (updateDraggable) {
+      this.stage.draggable(!this.canvasFitsInViewport())
+    }
+    
+    this.stage.batchDraw()
+  }
+  
+  // Handle wheel zoom with pointer-relative scaling
+  handleWheel(event, options = {}) {
+    const { updateDraggable = false, hitLayer = null } = options
+    
+    event.evt.preventDefault()
+    
+    if (hitLayer) hitLayer.listening(false)
+    
+    const oldScale = this.stage.scaleX() || 1
+    const pointer = this.stage.getPointerPosition()
+    // Convert screen coordinates to canvas coordinates before zoom
+    const canvasPoint = {
+      x: (pointer.x - this.stage.x()) / oldScale,
+      y: (pointer.y - this.stage.y()) / oldScale
+    }
+    
+    const direction = event.evt.deltaY > 0 ? 1 : -1
+    const nextScale = direction > 0 ? oldScale / SCALE_BY : oldScale * SCALE_BY
+    const minScale = this.getMinScale()
+    const maxScale = this.getMaxScale()
+    const newScale = clamp(nextScale, minScale, maxScale)
+    
+    // Zoom relative to pointer - keep canvas point under cursor
+    this.stage.scale({ x: newScale, y: newScale })
+    this.stage.position({
+      x: pointer.x - canvasPoint.x * newScale,
+      y: pointer.y - canvasPoint.y * newScale
+    })
+    
+    // Only center when zoomed out (content fits in viewport)
+    if (this.canvasFitsInViewport()) {
+      this.centerCanvas()
+    }
+    
+    if (updateDraggable) {
+      this.stage.draggable(!this.canvasFitsInViewport())
+    }
+    
+    this.stage.batchDraw()
+    
+    if (hitLayer) {
+      requestAnimationFrame(() => { hitLayer.listening(true) })
+    }
+  }
+  
+  // Handle touch pinch-zoom and pan
+  handleTouchMove(event) {
+    const touchOne = event.evt.touches[0]
+    const touchTwo = event.evt.touches[1]
+    
+    if (touchOne && touchTwo) {
+      event.evt.preventDefault()
+      this.stage.draggable(false)
+      
+      const pointOne = { x: touchOne.clientX, y: touchOne.clientY }
+      const pointTwo = { x: touchTwo.clientX, y: touchTwo.clientY }
+      const center = getCenter(pointOne, pointTwo)
+      const distance = getDistance(pointOne, pointTwo)
+      
+      if (!this.lastTouchCenter) {
+        this.lastTouchCenter = center
+        this.lastTouchDistance = distance
+        return
+      }
+      
+      const scale = this.stage.scaleX() * (distance / this.lastTouchDistance)
+      const minScale = this.getMinScale()
+      const maxScale = this.getMaxScale()
+      const clampedScale = clamp(scale, minScale, maxScale)
+      // Convert center point to canvas coordinates before zoom
+      const pointTo = {
+        x: (center.x - this.stage.x()) / this.stage.scaleX(),
+        y: (center.y - this.stage.y()) / this.stage.scaleY()
+      }
+      
+      this.stage.scale({ x: clampedScale, y: clampedScale })
+      
+      // Apply pan delta from touch movement
+      const dx = center.x - this.lastTouchCenter.x
+      const dy = center.y - this.lastTouchCenter.y
+      
+      this.stage.position({
+        x: center.x - pointTo.x * clampedScale + dx,
+        y: center.y - pointTo.y * clampedScale + dy
+      })
+      
+      this.constrainStageDrag()
+      
+      this.lastTouchCenter = center
+      this.lastTouchDistance = distance
+      this.stage.batchDraw()
+    }
+  }
+  
+  handleTouchEnd() {
+    this.lastTouchCenter = null
+    this.lastTouchDistance = 0
+  }
+  
   // Render group bounding boxes and labels (shared by all views)
   renderGroups() {
     const theme = this.theme
